Move code for rendering timeseries as JSON to separate file; simplify
