+++ lib/forecast_utils.js
@@ -16,60 +16,11 @@ const log = bunyan.createLogger({
 
 const cd2u = require('./cosmo-d2_utils')
 const tsCsv = require('./timeseries_as_csv')
+const tsJson = require('./timeseries_as_json')
 
 const secInHour = 3600
 const milliseconds = 1000
 const hoursInDay = 24
-
-
-function createDescriptionString (
-  vois,
-  stationId,
-  modelRun,
-  model,
-  startTimestamp,
-  endTimestamp
-) {
-  const showStartTimestamp = moment(startTimestamp).format('YYYY-MM-DDTHH:MM')
-  const showEndTimestamp = moment(endTimestamp).format('YYYY-MM-DDTHH:MM')
-  return `Forecast for quantities ${vois.join(
-    ', '
-  )} at station ${stationId} based on the ${modelRun} o'clock run of the ${model.toUpperCase()} model from ${showStartTimestamp} to ${showEndTimestamp}`
-}
-
-function renderJSONCosmoMosmix (
-  vois,
-  stationId,
-  modelRun,
-  model,
-  startTimestamp,
-  endTimestamp,
-  voiConfigs,
-  timeseriesDataArray
-) {
-  const descriptionString = createDescriptionString(
-    vois,
-    stationId,
-    modelRun,
-    model,
-    startTimestamp,
-    endTimestamp
-  )
-  const result = {
-    description: descriptionString,
-    data: []
-  }
-
-  _.forEach(voiConfigs, (voiConfig) => {
-    result.data.push({
-      label: voiConfig.target.key,
-      unit: voiConfig.target.unit,
-      timeseries: timeseriesDataArray[voiConfig.target.key]
-    })
-  })
-
-  return result
-}
 
 function deriveCsvFilePath (csvBasePath, type, timestamp, stationId, modelRun) {
   // No test for this function
@@ -189,7 +140,7 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
       functionToReadData: readMosmixTimeseriesData,
       MODEL_DATA_PATH: MOSMIX_DATA_BASE_PATH,
       unitsConverter: gu.convertUnitsForMosmix,
-      jsonRenderer: renderJSONCosmoMosmix,
+      jsonRenderer: tsJson.renderTimeseriesAsJSON,
       csvRenderer: tsCsv.renderTimeseriesAsCSV
     }
   } else if (model === 'cosmo-d2') {
@@ -199,7 +150,7 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
       functionToReadData: cd2u.readTimeseriesDataCosmoD2,
       MODEL_DATA_PATH: COSMO_DATA_BASE_PATH,
       unitsConverter: gu.convertUnitsForCosmo,
-      jsonRenderer: renderJSONCosmoMosmix,
+      jsonRenderer: tsJson.renderTimeseriesAsJSON,
       csvRenderer: tsCsv.renderTimeseriesAsCSV
     }
   } else {
@@ -207,7 +158,6 @@ function getConfigFromModel (model, COSMO_DATA_BASE_PATH, MOSMIX_DATA_BASE_PATH)
   }
 }
 
-exports.renderJSONCosmoMosmix = renderJSONCosmoMosmix
 exports.readMosmixTimeseriesData = readMosmixTimeseriesData
 exports.dropTimeseriesDataNotOfInteresWithParameter = dropTimeseriesDataNotOfInteresWithParameter
 exports.checkIfValidModelRun = checkIfValidModelRun
