+++ lib/measured_values_utils.js
@@ -66,28 +66,7 @@ function convertUnits (voiConfigs, timeseriesDataCollection) {
   return data
 }
 
-// Functions for building JSON-representation --------------------------------------- //
-
-function renderMeasuredValuesAsJSON (voiConfigs, timeseriesDataArray, vois, sid) {
-  const result = {
-    description: `Quantities ${vois} measured at station ${sid}`,
-    data: []
-  }
-
-  _.forEach(voiConfigs, (voiConfig) => {
-    result.data.push({
-      label: voiConfig.target.key,
-      unit: voiConfig.target.unit,
-      timeseries: timeseriesDataArray[voiConfig.target.key]
-    })
-  })
-
-  return result
-}
-
-
 // Export functions
-exports.renderMeasuredValuesAsJSON = renderMeasuredValuesAsJSON
 exports.dropTimeseriesDataNotOfInterest = dropTimeseriesDataNotOfInterest
 exports.dropNaN = dropNaN
 exports.convertUnits = convertUnits
